pyproject.toml
[tool.poetry]
name = "Hackaton"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
djangorestframework = "^3.15.1"
django-debug-toolbar = "^4.3.0"
django = "^5.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
